pkg: Rename cmd/ to promobot/
