+++ Makefile
@@ -38,7 +38,7 @@ test-mac:
 		//pkg/api/files:go_default_test \
 		//pkg/audit:go_default_test \
 		//pkg/dockerregistry:go_default_test \
-		//pkg/cmd:go_default_test \
+		//pkg/promobot:go_default_test \
 		//pkg/vulndash/adapter:go_default_test
 test-ci: download
 	make build
