+++ cmd/promobot-generate-manifest/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "sigs.k8s.io/k8s-container-image-promoter/cmd/promobot-generate-manifest",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/cmd:go_default_library",
+        "//pkg/promobot:go_default_library",
         "@io_k8s_klog//:go_default_library",
         "@io_k8s_sigs_yaml//:go_default_library",
         "@org_golang_x_xerrors//:go_default_library",
