+++ pkg/promobot/testdata/files-manifest.yaml
