+++ pkg/promobot/testdata/files/blue.png
