+++ pkg/promobot/testdata/files/green.png
