+++ pkg/promobot/testdata/files/red.png
