+++ pkg/promobot/testdata/manifests/manyfiles/expected.yaml
