+++ pkg/promobot/testdata/manifests/manyfiles/files/blue.yaml
