+++ pkg/promobot/testdata/manifests/manyfiles/files/green.yaml
