+++ pkg/promobot/testdata/manifests/manyfiles/files/red.yaml
