+++ pkg/promobot/testdata/manifests/manyfiles/filestores.yaml
