+++ pkg/promobot/testdata/manifests/onefiles/expected.yaml
