+++ pkg/promobot/testdata/manifests/onefiles/files.yaml
