+++ pkg/promobot/testdata/manifests/onefiles/filestores.yaml
